requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
django-anymail[mandrill]
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.11.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==41.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2.3
django-admin-honeypot @ https://github.com/dmpayton/django-admin-honeypot/archive/a840496d183df89965eeb33c9a5dd9ea3919dfff.zip
django-allauth==0.51.0
django-appconf==1.0.5
django-axes==5.39.0
django-bootstrap5==21.3
django-colorfield==0.7.2
django-compressor==4.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-csp==3.7
django-filter==22.1
django-ipware==4.0.2
django-jwt-oidc==1.0.11
django-libsass==0.9
django-password-validators==1.7.0
django-recaptcha==3.0.0
django-tables2==2.4.1
django-tex==1.1.10
djangorestframework==3.14.0
filetype==1.1.0
flake8==4.0.1
frozenlist==1.3.3
gunicorn
idna==3.3
Jinja2==3.1.2
jwcrypto==1.4
libsass==0.22.0
MarkupSafe==2.1.3
mccabe==0.6.1
multidict==6.0.3
oauthlib==3.2.2
Pillow==10.0.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pyflakes==2.4.0
PyJWT==2.6.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
rcssmin==1.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.16.0
slackclient==2.9.4
soupsieve==2.3.2.post1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
urllib3==1.26.18
uuid==1.30
wrapt==1.14.1
yarl==1.8.2
